requirements.txt
flask
requests
pysta>=0.0.30
gunicorn